+++ src/main/java/Duke.java
@@ -1,3 +1,6 @@
+import duke.*;
+import duke.command.Command;
+
 public class Duke {
     private Storage storage;
     private Ui ui;
