+++ src/main/java/duke/command/DeadlineAddCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class DeadlineAddCommand extends AddCommand {
     private String description;
     private String time;
