+++ src/main/java/duke/command/EventAddCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class EventAddCommand extends AddCommand {
     private String description;
     private String time;
