+++ src/main/java/duke/command/HelpCommand.java
@@ -1,12 +1,14 @@
-public class ListCommand implements Command {
+package duke.command;
 
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class HelpCommand implements Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        boolean allDone;
-        allDone = taskList.printItems();
-        if (allDone) {
-            ui.printCompleted();
-        }
+        ui.helpMessage();
     }
 
     @Override
