+++ src/main/java/duke/task/Event.java
@@ -1,4 +1,5 @@
-import java.time.LocalDate;
+package duke.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
